# frontend/src/main/resources/openapi.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "JurassicCRM",
    "termsOfService" : "Terms",
    "contact" : {
      "name" : "Dmitrii Fitisov",
      "url" : "https://google.com",
      "email" : "fitisovdmtr@gmail.com"
    }
  },
  "host" : "http://localhost",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/api/aviary/type" : {
      "get" : {
        "tags" : [ "aviaryType" ],
        "summary" : "getAllAviaries",
        "description" : "",
        "operationId" : "getAllAviaries",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SimpleEntityOutputTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "aviaryType" ],
        "summary" : "createAviary",
        "description" : "",
        "operationId" : "createAviary",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
    "/api/aviary/type/{id}" : {
      "put" : {
        "tags" : [ "aviaryType" ],
        "summary" : "updateAviary",
        "description" : "",
        "operationId" : "updateAviary",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "aviaryType" ],
        "summary" : "deleteAviary",
        "description" : "",
        "operationId" : "deleteAviary",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
...
    "/api/decoration/type/{id}" : {
      "put" : {
        "tags" : [ "decorationType" ],
        "summary" : "updateDecoration",
        "description" : "",
        "operationId" : "updateDecoration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "decorationType" ],
        "summary" : "deleteDecoration",
        "description" : "",
        "operationId" : "deleteDecoration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
...
    "/api/dinosaur/type/{id}" : {
      "put" : {
        "tags" : [ "dinosaur" ],
        "summary" : "updateDinosaurType",
        "description" : "",
        "operationId" : "updateDinosaurType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "dinosaur" ],
        "summary" : "deleteDinosaurType",
        "description" : "",
        "operationId" : "deleteDinosaurType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
...
    "/api/group/{groupId}/user/{userId}" : {
      "delete" : {
        "tags" : [ "group" ],
        "summary" : "removeUser",
        "description" : "",
        "operationId" : "removeUser",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/user" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "getUsers",
        "description" : "",
        "operationId" : "getUsers",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FullUserOutputTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "user" ],
        "summary" : "createUser",
        "description" : "",
        "operationId" : "createUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FullUserInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FullUserOutputTO"
            }
          }
        }
      }
    },
...
    "/api/user/role-all" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "findAllByRolesAll",
        "description" : "",
        "operationId" : "findAllByRolesAll",
        "parameters" : [ {
          "name" : "roles",
          "in" : "query",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "DOCUMENT_READER", "DINOSAUR_PASSPORT_READER", "AVIARY_PASSPORT_READER", "THEME_ZONE_PROJECT_READER", "TECHNOLOGICAL_MAP_READER", "RESEARCH_DATA_READER", "DOCUMENT_WRITER", "DINOSAUR_PASSPORT_WRITER", "AVIARY_PASSPORT_WRITER", "THEME_ZONE_PROJECT_WRITER", "TECHNOLOGICAL_MAP_WRITER", "RESEARCH_DATA_WRITER", "TASK_READER", "INCUBATION_TASK_READER", "AVIARY_BUILDING_TASK_READER", "RESEARCH_TASK_READER", "TASK_WRITER", "INCUBATION_TASK_WRITER", "AVIARY_BUILDING_TASK_WRITER", "RESEARCH_TASK_WRITER", "SECURITY_READER", "SECURITY_WRITER", "ADMIN" ]
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FullUserOutputTO"
              }
            }
          }
        }
      }
    },
    "/api/user/role-any" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "findAllByRolesAny",
        "description" : "",
        "operationId" : "findAllByRolesAny",
        "parameters" : [ {
          "name" : "roles",
          "in" : "query",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "DOCUMENT_READER", "DINOSAUR_PASSPORT_READER", "AVIARY_PASSPORT_READER", "THEME_ZONE_PROJECT_READER", "TECHNOLOGICAL_MAP_READER", "RESEARCH_DATA_READER", "DOCUMENT_WRITER", "DINOSAUR_PASSPORT_WRITER", "AVIARY_PASSPORT_WRITER", "THEME_ZONE_PROJECT_WRITER", "TECHNOLOGICAL_MAP_WRITER", "RESEARCH_DATA_WRITER", "TASK_READER", "INCUBATION_TASK_READER", "AVIARY_BUILDING_TASK_READER", "RESEARCH_TASK_READER", "TASK_WRITER", "INCUBATION_TASK_WRITER", "AVIARY_BUILDING_TASK_WRITER", "RESEARCH_TASK_WRITER", "SECURITY_READER", "SECURITY_WRITER", "ADMIN" ]
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/FullUserOutputTO"
              }
            }
          }
        }
      }
    },
    "/api/user/{userId}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "getUserById",
        "description" : "",
        "operationId" : "getUserById",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FullUserOutputTO"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "user" ],
        "summary" : "updateUser",
        "description" : "",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/FullUserInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FullUserOutputTO"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "FullUserInputTO" : {
      "type" : "object",
      "required" : [ "department" ],
      "properties" : {
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "department" : {
          "type" : "string",
          "enum" : [ "RESEARCH", "INCUBATION", "SECURITY", "ADMINISTRATION", "MAINTENANCE", "ACCOMMODATION" ]
        },
        "groupIds" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    },
    "FullUserOutputTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "department" : {
          "type" : "string",
          "enum" : [ "RESEARCH", "INCUBATION", "SECURITY", "ADMINISTRATION", "MAINTENANCE", "ACCOMMODATION" ]
        },
        "groupIds" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SimpleEntityOutputTO"
          }
        }
      }
    },
...
    "SimpleEntityInputTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
    "SimpleEntityOutputTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
  }
}